employer and developer models
